Renamed KayakTest to NancyInGate
