--- KayakTest/Models/UserModel.cs
@@ -1,9 +1,0 @@
-﻿namespace NancyInGate.Models
-{
-    public class UserModel
-    {
-        public string FirstName { get; set; }
-
-        public string LastName { get; set; }
-    }
-}
