+++ NancyInGate/Models/UserModel.cs
@@ -1,0 +1,9 @@
+﻿namespace NancyInGate.Models
+{
+    public class UserModel
+    {
+        public string FirstName { get; set; }
+
+        public string LastName { get; set; }
+    }
+}
