+++ NancyInGate/NancyInGate.csproj
