+++ NancyInGate/Views/test.cshtml
@@ -1,0 +1,9 @@
+﻿<html>
+<head>
+    <title>Nancy in a Kayak</title>
+</head>
+<body>
+    <h1>Hello there @Model.FirstName @Model.LastName!</h1>
+    <p>Razor and Nancy in a Kayak!</p>
+</body>
+</html>
