+++ NancyInGate/app.manifest
