+++ NancyInGate/packages.config
@@ -1,0 +1,19 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Firefly" version="0.4" />
+  <package id="Gate" version="0.3.6" />
+  <package id="Gate.Adapters.Nancy" version="0.3.6" />
+  <package id="Gate.Hosts.Firefly" version="0.3.6" />
+  <package id="Gate.Hosts.HttpListener" version="0.3.6" />
+  <package id="Gate.Hosts.Kayak" version="0.3.6" />
+  <package id="Gate.Hosts.Manos" version="0.3.6" />
+  <package id="Gate.Middleware" version="0.3.6" />
+  <package id="Gate.Stack.HttpListener.Nancy" version="0.3.2" />
+  <package id="Gate.Stack.Kayak.Nancy" version="0.3.2" />
+  <package id="Kayak" version="0.7.2" />
+  <package id="Nancy" version="0.11.0" />
+  <package id="Nancy.Hosting.Owin" version="0.11.0" />
+  <package id="Nancy.Viewengines.Razor" version="0.11.0" />
+  <package id="Owin" version="0.7" />
+  <package id="Spark" version="1.7.2.0" />
+</packages>
